requirements.txt
Django
psycopg2
youtube_transcript_api
validators
dj-database-url
djangorestframework